Updated CMakeLists.txt to work in Labs
I fixed up the CMakeLists to use CMake variables and preserve normal
functionality so the CMAKE_PREFIX_PATH uses the Macports and ENCS paths
to change search functions. Comments were added as well. This should fix
the GLEW target on OSX and Linux.

Similarly, the ENCS Eigen3 installation is setup to work using CMake's
functionality and should work in other environments as well.

GLM isn't properly installed in ENCS so the include directory has to be
added directly.